user search ability for both topic and all articles
