+++ src/Components/Articles-Topic/TopicArticles.jsx
@@ -1,7 +1,8 @@
 import { ArticlesCard } from "../Cards/ArticlesCard";
 import { useEffect, useState } from "react";
 import * as API from '../../Api'
-import { Link } from 'react-router-dom'
+import { Link, useSearchParams } from 'react-router-dom'
+import { searchArticles } from "../../Utils/Utils";
 
 
 
@@ -9,6 +10,8 @@ export const TopicArticles = ({topic}) => {
     const [articles, setArticles] = useState([]);
     const [isLoading, setIsLoading] = useState(true);
     const [err, setErr] = useState(null);
+    const [searchParams, setSearchParams] = useSearchParams();
+    const param = searchParams.get('search')
 
     useEffect(() => {
         setIsLoading(true);
@@ -19,20 +22,26 @@ export const TopicArticles = ({topic}) => {
         })
     },[topic])
 
+    const searchedArticles = searchArticles(param, articles)
+
     if(isLoading) return <h2>Loading ...</h2>
     else
     if(err) return <h2>{err}</h2>
     else
     return (
-    <ul className="all-articles">
-        {
-        articles.map(article => {
-            return (
-                <Link to={`/article/${article.article_id}`}>
+        <ul className="all-articles">
+            {param ?
+            searchedArticles.map(article => {
+                return <Link to={`/article/${article.article_id}`}>
                     <ArticlesCard className="article-card" key={article.article_id} article={article}/>
                 </Link>
-            )
-    })}
-    </ul>
+    
+        }): articles.map(article => {
+            return <Link to={`/article/${article.article_id}`}>
+                <ArticlesCard className="article-card" key={article.article_id} article={article}/>
+            </Link>
+        })}
+    
+        </ul>
     )
 }
